+++ server/src/main/java/com/unihub/app/repositories/community/content/FolderRepository.java
@@ -11,10 +11,10 @@ import java.util.UUID;
 public interface FolderRepository extends JpaRepository<Folder, UUID> {
 
     @Query("SELECT f FROM Folder f LEFT JOIN FETCH f.owner WHERE f.course.id = :courseId AND f.parentFolder.id = :parentFolderId ORDER BY f.name ASC")
-    List<Folder> findByCourseIdAndParentFolderId(@Param("courseId") int courseId, @Param("parentFolderId") UUID parentFolderId);
+    List<Folder> findByCourseIdAndParentFolderId(@Param("courseId") Long courseId, @Param("parentFolderId") UUID parentFolderId);
 
     @Query("SELECT f FROM Folder f LEFT JOIN FETCH f.owner WHERE f.course.id = :courseId AND f.parentFolder IS NULL ORDER BY f.name ASC")
-    List<Folder> findRootFoldersByCourseId(@Param("courseId") int courseId);
+    List<Folder> findRootFoldersByCourseId(@Param("courseId") Long courseId);
 
-    boolean existsByIdAndCourseId(UUID id, int courseId);
+    boolean existsByIdAndCourseId(UUID id, Long courseId);
 }
